Create enterstocks.cpp
Program to input names of stocks and make a first analysis of curves about the Thinkorswim trading platform graphs displayed on screen to choose the better stocks to invest. This first version read a column of pixels to analyze the Stochastic curve. Next version will read more columns of pixels to analyze the trend of mobile average curves and MACD parameters.